+++ packages/sdk/src/lib/dataProtectorSharing/buyProtectedData.ts
@@ -1,3 +1,4 @@
+import { AddressLike, BigNumberish, ContractTransactionResponse } from 'ethers';
 import { WorkflowError } from '../../utils/errors.js';
 import { resolveENS } from '../../utils/resolveENS.js';
 import {
@@ -12,9 +13,12 @@ import {
   SuccessWithTransactionHash,
 } from '../types/index.js';
 import { IExecConsumer } from '../types/internalTypes.js';
-import { approveCollectionContract } from './smartContract/approveCollectionContract.js';
+import { approveProtectedDataForCollectionContract } from './smartContract/approveProtectedDataForCollectionContract.js';
+import { getPocoContract } from './smartContract/getPocoContract.js';
 import { getSharingContract } from './smartContract/getSharingContract.js';
+import { getAccountDetails } from './smartContract/pocoContract.reads.js';
 import {
+  onlyAccountWithMinimumBalance,
   onlyCollectionOperator,
   onlyProtectedDataCurrentlyForSale,
   onlyValidSellingParams,
@@ -52,25 +56,57 @@ export async function buyProtectedData({
   let userAddress = await iexec.wallet.getAddress();
   userAddress = userAddress.toLowerCase();
 
-  const sharingContract = await getSharingContract(
-    iexec,
-    sharingContractAddress
-  );
+  const [sharingContract, pocoContract] = await Promise.all([
+    getSharingContract(iexec, sharingContractAddress),
+    getPocoContract(iexec),
+  ]);
 
   //---------- Smart Contract Call ----------
-  const protectedDataDetails = await getProtectedDataDetails({
-    sharingContract,
-    protectedData: vProtectedData,
-    userAddress,
-  });
+  const [protectedDataDetails, accountDetails] = await Promise.all([
+    getProtectedDataDetails({
+      sharingContract,
+      protectedData: vProtectedData,
+      userAddress,
+    }),
+    getAccountDetails({
+      pocoContract,
+      userAddress,
+      sharingContractAddress,
+    }),
+  ]);
 
   //---------- Pre flight check----------
   onlyProtectedDataCurrentlyForSale(protectedDataDetails);
-  onlyValidSellingParams(price, protectedDataDetails.sellingParams.price);
+  onlyValidSellingParams(vPrice, protectedDataDetails.sellingParams.price);
+  onlyAccountWithMinimumBalance({
+    accountDetails,
+    minimumBalance: vPrice,
+  });
 
   try {
-    let tx;
     const { txOptions } = await iexec.config.resolveContractsClient();
+
+    let tx: ContractTransactionResponse;
+    const buyProtectedDataCallParams: [AddressLike, AddressLike, BigNumberish] =
+      [vProtectedData, userAddress, vPrice];
+    if (accountDetails.sharingContractAllowance >= BigInt(vPrice)) {
+      tx = await sharingContract.buyProtectedData(
+        ...buyProtectedDataCallParams,
+        txOptions
+      );
+    } else {
+      const callData = sharingContract.interface.encodeFunctionData(
+        'buyProtectedData',
+        buyProtectedDataCallParams
+      );
+      tx = await pocoContract.approveAndCall(
+        sharingContractAddress,
+        vPrice,
+        callData,
+        txOptions
+      );
+    }
+    await tx.wait();
 
     if (vAddToCollectionId) {
       await onlyCollectionOperator({
@@ -79,16 +115,8 @@ export async function buyProtectedData({
         userAddress,
       });
 
-      // should implement multicall in the future
-      tx = await sharingContract.buyProtectedData(
-        vProtectedData,
-        userAddress,
-        vPrice,
-        txOptions
-      );
-      await tx.wait();
-
-      await approveCollectionContract({
+      // should implement multicall in the future (not compatible with approveAndCall)
+      await approveProtectedDataForCollectionContract({
         iexec,
         protectedData: vProtectedData,
         sharingContractAddress,
@@ -102,14 +130,6 @@ export async function buyProtectedData({
           txOptions
         );
       await txAddToCollection.wait();
-    } else {
-      tx = await sharingContract.buyProtectedData(
-        vProtectedData,
-        userAddress,
-        vPrice,
-        txOptions
-      );
-      await tx.wait();
     }
 
     return {